// ui/src/interfaces.ts
export enum UserOperation {
  Login,
  Register,
  CreateCommunity,
  CreatePost,
  ListCommunities,
  ListCategories,
  GetPost,
  GetCommunity,
  CreateComment,
  EditComment,
  SaveComment,
  CreateCommentLike,
  GetPosts,
  CreatePostLike,
  CreatePostTagged,
  EditPost,
  SavePost,
  EditCommunity,
  FollowCommunity,
  GetFollowedCommunities,
  GetUserDetails,
  GetReplies,
  GetUserMentions,
  EditUserMention,
  GetModlog,
  BanFromCommunity,
  AddModToCommunity,
  CreateSite,
  EditSite,
  GetSite,
  AddAdmin,
  BanUser,
  Search,
  MarkAllAsRead,
  SaveUserSettings,
  TransferCommunity,
  TransferSite,
  DeleteAccount,
  PasswordReset,
  PasswordChange,
  CreatePrivateMessage,
  EditPrivateMessage,
  GetPrivateMessages,
  UserJoin,
  GetComments,
  GetSiteConfig,
  SaveSiteConfig,
}
// ...
export interface GetFollowedCommunitiesResponse {
  communities: Array<CommunityUser>;
}
// ...
export interface GetRepliesResponse {
  replies: Array<Comment>;
}

export interface GetUserMentionsForm {
  sort: string;
  page?: number;
  limit?: number;
  unread_only: boolean;
  auth?: string;
}

export interface GetUserMentionsResponse {
  mentions: Array<Comment>;
}

export interface EditUserMentionForm {
  user_mention_id: number;
  read?: boolean;
  auth?: string;
}

export interface UserMentionResponse {
  mention: Comment;
}

export interface BanFromCommunityForm {
  community_id: number;
  user_id: number;
  ban: boolean;
  reason?: string;
  expires?: number;
  auth?: string;
}

export interface BanFromCommunityResponse {
  user: UserView;
  banned: boolean;
}

export interface AddModToCommunityForm {
  community_id: number;
  user_id: number;
  added: boolean;
  auth?: string;
}

export interface TransferCommunityForm {
  community_id: number;
  user_id: number;
  auth?: string;
}

export interface TransferSiteForm {
  user_id: number;
  auth?: string;
}

export interface AddModToCommunityResponse {
  moderators: Array<CommunityUser>;
}

export interface GetModlogForm {
  mod_user_id?: number;
  community_id?: number;
  page?: number;
  limit?: number;
}

export interface GetModlogResponse {
  removed_posts: Array<ModRemovePost>;
  locked_posts: Array<ModLockPost>;
  stickied_posts: Array<ModStickyPost>;
  removed_comments: Array<ModRemoveComment>;
  removed_communities: Array<ModRemoveCommunity>;
  banned_from_community: Array<ModBanFromCommunity>;
  banned: Array<ModBan>;
  added_to_community: Array<ModAddCommunity>;
  added: Array<ModAdd>;
}

export interface ModRemovePost {
  id: number;
  mod_user_id: number;
  post_id: number;
  reason?: string;
  removed?: boolean;
  when_: string;
  mod_user_name: string;
  post_name: string;
  community_id: number;
  community_name: string;
}

export interface ModLockPost {
  id: number;
  mod_user_id: number;
  post_id: number;
  locked?: boolean;
  when_: string;
  mod_user_name: string;
  post_name: string;
  community_id: number;
  community_name: string;
}

export interface ModStickyPost {
  id: number;
  mod_user_id: number;
  post_id: number;
  stickied?: boolean;
  when_: string;
  mod_user_name: string;
  post_name: string;
  community_id: number;
  community_name: string;
}

export interface ModRemoveComment {
  id: number;
  mod_user_id: number;
  comment_id: number;
  reason?: string;
  removed?: boolean;
  when_: string;
  mod_user_name: string;
  comment_user_id: number;
  comment_user_name: string;
  comment_content: string;
  post_id: number;
  post_name: string;
  community_id: number;
  community_name: string;
}

export interface ModRemoveCommunity {
  id: number;
  mod_user_id: number;
  community_id: number;
  reason?: string;
  removed?: boolean;
  expires?: number;
  when_: string;
  mod_user_name: string;
  community_name: string;
}

export interface ModBanFromCommunity {
  id: number;
  mod_user_id: number;
  other_user_id: number;
  community_id: number;
  reason?: string;
  banned?: boolean;
  expires?: number;
  when_: string;
  mod_user_name: string;
  other_user_name: string;
  community_name: string;
}

export interface ModBan {
  id: number;
  mod_user_id: number;
  other_user_id: number;
  reason?: string;
  banned?: boolean;
  expires?: number;
  when_: string;
  mod_user_name: string;
  other_user_name: string;
}

export interface ModAddCommunity {
  id: number;
  mod_user_id: number;
  other_user_id: number;
  community_id: number;
  removed?: boolean;
  when_: string;
  mod_user_name: string;
  other_user_name: string;
  community_name: string;
}

export interface ModAdd {
  id: number;
  mod_user_id: number;
  other_user_id: number;
  removed?: boolean;
  when_: string;
  mod_user_name: string;
  other_user_name: string;
}

export interface LoginForm {
  username_or_email: string;
  password: string;
}

export interface RegisterForm {
  username: string;
  email?: string;
  password: string;
  password_verify: string;
  admin: boolean;
  show_nsfw: boolean;
}

export interface LoginResponse {
  jwt: string;
}
// ...
export interface CommunityResponse {
  community: Community;
}

export interface ListCommunitiesForm {
  sort: string;
  page?: number;
  limit?: number;
  auth?: string;
}

export interface ListCommunitiesResponse {
  communities: Array<Community>;
}

export interface ListCategoriesResponse {
  categories: Array<Category>;
}
// ...
export interface PostResponse {
  post: Post;
}

export interface CommentForm {
  content: string;
  post_id: number;
  parent_id?: number;
  edit_id?: number;
  creator_id: number;
  removed?: boolean;
  deleted?: boolean;
  reason?: string;
  read?: boolean;
  auth: string;
}

export interface SaveCommentForm {
  comment_id: number;
  save: boolean;
  auth?: string;
}

export interface CommentResponse {
  comment: Comment;
  recipient_ids: Array<number>;
}

export interface CommentLikeForm {
  comment_id: number;
  post_id: number;
  score: number;
  auth?: string;
}

export interface CommentNode {
  comment: Comment;
  children?: Array<CommentNode>;
}

export interface GetPostsForm {
  type_: string;
  sort: string;
  page?: number;
  limit?: number;
  community_id?: number;
  auth?: string;
}

export interface GetPostsResponse {
  posts: Array<Post>;
}
// ...
export interface GetSiteConfig {
  auth?: string;
}

export interface GetSiteConfigResponse {
  config_hjson: string;
}

export interface SiteConfigForm {
  config_hjson: string;
  auth?: string;
}

export interface GetSiteResponse {
  site: Site;
  admins: Array<UserView>;
  banned: Array<UserView>;
  online: number;
}

export interface SiteResponse {
  site: Site;
}

export interface BanUserForm {
  user_id: number;
  ban: boolean;
  reason?: string;
  expires?: number;
  auth?: string;
}

export interface BanUserResponse {
  user: UserView;
  banned: boolean;
}

export interface AddAdminForm {
  user_id: number;
  added: boolean;
  auth?: string;
}

export interface AddAdminResponse {
  admins: Array<UserView>;
}

export interface SearchForm {
  q: string;
  type_: string;
  community_id?: number;
  sort: string;
  page?: number;
  limit?: number;
  auth?: string;
}

export interface SearchResponse {
  type_: string;
  posts?: Array<Post>;
  comments?: Array<Comment>;
  communities: Array<Community>;
  users: Array<UserView>;
}
// ...
export interface PrivateMessagesResponse {
  messages: Array<PrivateMessage>;
}

export interface PrivateMessageResponse {
  message: PrivateMessage;
}
// ...
type ResponseType =
  | SiteResponse
  | GetFollowedCommunitiesResponse
  | ListCommunitiesResponse
  | GetPostsResponse
  | PostResponse
  | GetRepliesResponse
  | GetUserMentionsResponse
  | ListCategoriesResponse
  | CommunityResponse
  | CommentResponse
  | UserMentionResponse
  | LoginResponse
  | GetModlogResponse
  | SearchResponse
  | BanFromCommunityResponse
  | AddModToCommunityResponse
  | BanUserResponse
  | AddAdminResponse
  | PrivateMessageResponse
  | PrivateMessagesResponse
  | GetSiteConfigResponse;

export interface WebSocketResponse {
  op: UserOperation;
  data: ResponseType;
}

export interface WebSocketJsonResponse {
  op?: string;
  data?: ResponseType;
  error?: string;
  reconnect?: boolean;
}